Enhance database migration and development workflow
- Improve migration script with robust error handling and logging
- Add dynamic migration file detection and processing
- Update Docker Compose and development scripts to support migration workflow
- Modify database Dockerfile to include migration watcher
- Refactor migration error handling to continue processing on individual migration failures
- Add fallback logo rendering in frontend quote component

+++ backend/utils/security.go
@@ -24,8 +24,8 @@ func GetTicker(conn *Conn, securityId int, timestamp time.Time) (string, error) 
 	return ticker, nil
 
 }
-func GetCIKFromTicker(conn *Conn, ticker string, timestamp time.Time) (int, error) {
-	var cik int
+func GetCIKFromTicker(conn *Conn, ticker string, timestamp time.Time) (int64, error) {
+	var cik int64
 	err := conn.DB.QueryRow(context.Background(), "SELECT cik from securities where ticker = $1 and minDate <= $2 and (maxDate >= $2 or maxDate is NULL)", ticker, timestamp).Scan(&cik)
 	if err != nil {
 		return 0, fmt.Errorf("3333w0ngb %v", err)
